Moved some utility methods to a separate file so that they can be reused
