+++ tests/test_executor.py
@@ -1,37 +1,11 @@
-from datetime import timedelta
+import uuid
 from pathlib import Path
-from typing import Optional
 
-from psij import SubmitException, Job, JobExecutor, JobSpec, JobState, JobAttributes, JobStatus
+from psij import SubmitException, Job, JobSpec, JobState
 from tempfile import TemporaryDirectory
 
 from executor_test_params import ExecutorTestParams
-
-
-_QUICK_EXECUTORS = set(['local', 'batch-test'])
-
-
-def _make_test_dir() -> None:
-    (Path.home() / '.psij' / 'test').mkdir(parents=True, exist_ok=True)
-
-
-def _get_timeout(execparams: ExecutorTestParams) -> Optional[timedelta]:
-    if execparams.executor in _QUICK_EXECUTORS:
-        return timedelta(minutes=10)
-    else:
-        return None
-
-
-def assert_completed(status: Optional[JobStatus]) -> None:
-    assert status is not None
-    assert status.state == JobState.COMPLETED
-
-
-def _get_executor_instance(ep: ExecutorTestParams, job: Job) -> JobExecutor:
-    assert job.spec is not None
-    job.spec.launcher = ep.launcher
-    job.spec.attributes = JobAttributes(custom_attributes=ep.custom_attributes)
-    return JobExecutor.get_instance(ep.executor, url=ep.url)
+from _test_tools import _get_executor_instance, _get_timeout, assert_completed, _make_test_dir
 
 
 def test_simple_job(execparams: ExecutorTestParams) -> None:
